refactor (ordertable) : validator 구현체를 order 쪽으로 이동
